pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "csc"
version = "0.1.0"
description = "Clinical Supply Chain Multi-Agent Modeling System"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.42.0",
    "pydantic>=2.10.0",
    "click>=8.1.0",
    "rich>=13.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8.0",
]

[project.scripts]
csc = "csc.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/csc"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
